pyproject.toml
[project]
name = "jobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2pdf>=0.1.8",
]

[dependency-groups]
dev = [
    "ipython>=9.6.0",
]

[tool.uv]
required-environments=["sys_platform == 'darwin' and platform_machine == 'x86_64'"]
